feat: refactor memoItemModel
