+++ MemoFlux/MemoFlux/Models/MemoItemModel.swift
@@ -8,6 +8,7 @@
 import Foundation
 import SwiftData
 import SwiftUI
+import UIKit
 
 /// 首页 List cell 中 item 的数据模型
 @Model
@@ -20,7 +21,7 @@ final class MemoItemModel: Identifiable {
   var tags: [String] = []
   var createdAt: Date = Date()
   var scheduledDate: Date?
-  var source: String = ""
+  var source: String = ""  // 数据来源（快捷指令/手动添加）
   
   // API响应相关字段
   var apiResponseData: Data?  // 存储API响应的JSON数据
@@ -36,9 +37,19 @@ final class MemoItemModel: Identifiable {
   }
   
   // 获取API响应
-  var apiResponse: APIResponse? {
+  var apiResponse: MemoItemModel.APIResponse? {
     guard let apiResponseData = apiResponseData else { return nil }
-    return try? JSONDecoder().decode(APIResponse.self, from: apiResponseData)
+    return try? JSONDecoder().decode(MemoItemModel.APIResponse.self, from: apiResponseData)
+  }
+  
+  // 获取Information响应
+  var information: Information? {
+    return apiResponse?.information
+  }
+  
+  // 获取Schedule响应
+  var schedule: Schedule? {
+    return apiResponse?.schedule
   }
   
   init(image: UIImage, title: String = "", tags: [String], source: String = "") {
@@ -88,7 +99,7 @@ final class MemoItemModel: Identifiable {
   }
   
   /// API响应
-  func setAPIResponse(_ response: APIResponse) {
+  func setAPIResponse(_ response: MemoItemModel.APIResponse) {
     do {
       self.apiResponseData = try JSONEncoder().encode(response)
       self.hasAPIResponse = true
@@ -159,4 +170,182 @@ final class MemoItemModel: Identifiable {
   func removeTag(_ tag: String) {
     self.tags.removeAll { $0 == tag }
   }
+  
+  // MARK: - 嵌套模型
+  
+  // MARK: - 主响应模型
+  struct APIResponse: Codable, Equatable {
+    let mostPossibleCategory: String
+    let information: Information
+    let schedule: Schedule
+    
+    enum CodingKeys: String, CodingKey {
+      case mostPossibleCategory
+      case information
+      case schedule
+    }
+    
+    /// 获取所有任务的标签
+    var allTags: Set<String> {
+      var tags = Set<String>()
+      tags.formUnion(information.tags)
+      schedule.tasks.forEach { task in
+        tags.formUnion(task.tags)
+      }
+      return tags
+    }
+    
+    /// 获取所有任务
+    var allTasks: [ScheduleTask] {
+      return schedule.tasks
+    }
+    
+    /// 根据日期筛选任务
+    func tasks(for date: Date) -> [ScheduleTask] {
+      let calendar = Calendar.current
+      return schedule.tasks.filter { task in
+        guard let taskDate = task.startDate else { return false }
+        return calendar.isDate(taskDate, inSameDayAs: date)
+      }
+    }
+  }
+  
+  // MARK: - Information 模型
+  struct Information: Codable, Equatable {
+    let title: String
+    let informationItems: [InformationItem]
+    let relatedItems: [String]
+    let summary: String
+    let tags: [String]
+    
+    enum CodingKeys: String, CodingKey {
+      case title
+      case informationItems = "informationItems"
+      case relatedItems = "relatedItems"
+      case summary
+      case tags
+    }
+  }
+  
+  struct InformationItem: Codable, Identifiable, Equatable {
+    let id: Int
+    let header: String
+    let content: String
+    let node: InformationNode?
+  }
+  
+  struct InformationNode: Codable, Equatable {
+    let targetId: Int
+    let relationship: String
+    
+    enum CodingKeys: String, CodingKey {
+      case targetId
+      case relationship
+    }
+  }
+  
+  // MARK: - Schedule 模型
+  struct Schedule: Codable, Equatable {
+    let title: String
+    let category: String
+    let tasks: [ScheduleTask]
+  }
+  
+  struct ScheduleTask: Codable, Identifiable, Equatable {
+    let startTime: String
+    let endTime: String
+    let people: [String]
+    let theme: String
+    let coreTasks: [String]
+    let position: [String]
+    let tags: [String]
+    let category: String
+    let suggestedActions: [String]
+    let id: UUID
+    
+    enum CodingKeys: String, CodingKey {
+      case startTime = "startTime"
+      case endTime = "endTime"
+      case people
+      case theme
+      case coreTasks = "coreTasks"
+      case position
+      case tags
+      case category
+      case suggestedActions = "suggestedActions"
+      // 注意：id不包含在CodingKeys中，因为我们不从API解码它
+    }
+    
+    // 自定义初始化器，自动生成UUID
+    init(from decoder: Decoder) throws {
+      let container = try decoder.container(keyedBy: CodingKeys.self)
+      startTime = try container.decode(String.self, forKey: .startTime)
+      endTime = try container.decode(String.self, forKey: .endTime)
+      people = try container.decode([String].self, forKey: .people)
+      theme = try container.decode(String.self, forKey: .theme)
+      coreTasks = try container.decode([String].self, forKey: .coreTasks)
+      position = try container.decode([String].self, forKey: .position)
+      tags = try container.decode([String].self, forKey: .tags)
+      category = try container.decode(String.self, forKey: .category)
+      suggestedActions = try container.decode([String].self, forKey: .suggestedActions)
+      // 自动生成UUID
+      id = UUID()
+    }
+    
+    // 编码时不包含id字段
+    func encode(to encoder: Encoder) throws {
+      var container = encoder.container(keyedBy: CodingKeys.self)
+      try container.encode(startTime, forKey: .startTime)
+      try container.encode(endTime, forKey: .endTime)
+      try container.encode(people, forKey: .people)
+      try container.encode(theme, forKey: .theme)
+      try container.encode(coreTasks, forKey: .coreTasks)
+      try container.encode(position, forKey: .position)
+      try container.encode(tags, forKey: .tags)
+      try container.encode(category, forKey: .category)
+      try container.encode(suggestedActions, forKey: .suggestedActions)
+    }
+    
+    // 手动初始化器（用于测试和预览）
+    init(
+      startTime: String,
+      endTime: String,
+      people: [String],
+      theme: String,
+      coreTasks: [String],
+      position: [String],
+      tags: [String],
+      category: String,
+      suggestedActions: [String],
+      id: UUID = UUID()
+    ) {
+      self.startTime = startTime
+      self.endTime = endTime
+      self.people = people
+      self.theme = theme
+      self.coreTasks = coreTasks
+      self.position = position
+      self.tags = tags
+      self.category = category
+      self.suggestedActions = suggestedActions
+      self.id = id
+    }
+    
+    // 将字符串时间转换为Date
+    var startDate: Date? {
+      return ISO8601DateFormatter().date(from: startTime)
+    }
+    
+    var endDate: Date? {
+      return ISO8601DateFormatter().date(from: endTime)
+    }
+  }
 }
+
+// MARK: - 类型别名
+typealias APIResponse = MemoItemModel.APIResponse
+typealias Information = MemoItemModel.Information
+typealias InformationItem = MemoItemModel.InformationItem
+typealias InformationNode = MemoItemModel.InformationNode
+typealias Schedule = MemoItemModel.Schedule
+typealias ScheduleTask = MemoItemModel.ScheduleTask
