+++ MemoFlux/MemoFlux/Views/ListViews/IntentDiscoveryView.swift
@@ -187,7 +187,7 @@ private func createTestMemoItemsWithSchedule() -> [MemoItemModel] {
   )
   
   // 模拟 API 响应
-  let scheduleTask = ScheduleTask(
+  let scheduleTask = MemoItemModel.ScheduleTask(
     startTime: "2024-05-16T10:30:00+08:00",
     endTime: "2024-05-16T11:30:00+08:00",
     people: ["张三", "李四"],
@@ -200,16 +200,16 @@ private func createTestMemoItemsWithSchedule() -> [MemoItemModel] {
     id: UUID()
   )
   
-  let apiResponse = APIResponse(
+  let apiResponse = MemoItemModel.APIResponse(
     mostPossibleCategory: "schedule",
-    information: InformationResponse(
+    information: MemoItemModel.Information(
       title: "会议安排知识",
       informationItems: [],
       relatedItems: [],
       summary: "",
       tags: ["会议", "项目"]
     ),
-    schedule: ScheduleResponse(
+    schedule: MemoItemModel.Schedule(
       title: "今日日程",
       category: "工作",
       tasks: [scheduleTask]
